+++ src/tvdeblur.h
@@ -33,22 +33,22 @@ public:
     float lambda = 1e-3f; //! (initial) weight of TV penalty
     float min_lambda = 1e-5f; //! (final) weight of TV penalty (after annealing)
     float annealing_rate = 0.999f; //! rate by which lambda decreases
-    int max_inner_iterations = 1; //! max number of iterations
-    int max_anneal_rounds = 10000; //! max number of annealing rounds
-    int max_nonblind_iterations = 100; //! maximum number of iterations used in NONBLIND step
+    size_t max_inner_iterations = 1; //! max number of iterations
+    size_t max_anneal_rounds = 10000; //! max number of annealing rounds
+    size_t max_nonblind_iterations = 100; //! maximum number of iterations used in NONBLIND step
     float grad_descent_step_u = 5e-3f; //! step length in gradient descent (over image) 
     float grad_descent_step_k = 1e-3f; //! step length in gradient descent (over kernel)
     float dfx_tolerance = 1e-4f; //! smallest ( f(x)-f(x_+) )/f(x), stopping criterion
-    int ker_size = 9; // kernel width/height 
+    size_t ker_size = 9; // kernel width/height 
     int verbosity = 100;
     // pyramid parameters
     int pyramid_numlevels = 1;
     float pyramid_imsize_scaling = 1; // how much to resize the image at each time? (<1)
     float pyramid_lambda_scaling = 1; // how much to increase lambda at each time (>1)
     // other parameters
-    int rows = 0;
-    int cols = 0;
-    int chan = 0;
+    size_t rows = 0;
+    size_t cols = 0;
+    size_t chan = 0;
     bool vis_results = true; // write out images at every step?
     int kernel_init_type = KERNEL_INIT_FLAT;
 
@@ -62,9 +62,9 @@ class tvdeblur {
     enum truncation_type { TRUNCATION_FULL, TRUNCATION_SAME, TRUNCATION_VALID };
 public:
     struct pyramid_size {
-        int rows, cols;
-        int rows_ker, cols_ker;
-        int rows_padded, cols_padded;
+        size_t rows, cols;
+        size_t rows_ker, cols_ker;
+        size_t rows_padded, cols_padded;
         float lambda;
     }; // structure for multiscale processing
 
@@ -82,7 +82,7 @@ public:
     void solve_inner(float lambda);
     // for a given kernel, solve the non-blind deconvolution problem
     // (just using gradient descent)
-    void solve_nonblind(const vector<float>& k, float tv_penalty, int num_iters);
+    void solve_nonblind(const vector<float>& k, float tv_penalty, size_t num_iters);
     void solve();
     void debug();
 
@@ -101,9 +101,9 @@ public:
     void project_kernel_onto_feasible_set(vector<float>& k);
 
     vector<float> pad_image(const vector<float>& in, 
-                        int rows_in, int cols_in, int rows_out, int cols_out);
+                        size_t rows_in, size_t cols_in, size_t rows_out, size_t cols_out);
     vector<float> unpad_image(const vector<float>& in, 
-                        int rows_in, int cols_in, int rows_out, int cols_out);
+                        size_t rows_in, size_t cols_in, size_t rows_out, size_t cols_out);
 
     vector<float> fft_imfilter( const vector<float>& im1, int rows1, int cols1, int chan1, 
                                 const vector<float>& im2, int rows2, int cols2, int chan2,
@@ -124,9 +124,9 @@ public:
     float* ker_p1;
     fftwf_complex* ker_p2;
     // data sizes:
-    int rows, cols, chan;
-    int rows_ker, cols_ker;
-    int rows_padded, cols_padded;
+    size_t rows, cols, chan;
+    size_t rows_ker, cols_ker;
+    size_t rows_padded, cols_padded;
 };
 
 } // namespace bcv
